+++ test/server/index.js
@@ -15,13 +15,24 @@
 let SuperTokens = require("supertokens-node");
 let SuperTokensRaw = require("supertokens-node/lib/build/supertokens").default;
 let SessionRecipeRaw = require("supertokens-node/lib/build/recipe/session/recipe").default;
+let Querier = require("supertokens-node/lib/build/querier").Querier;
+let NormalisedURLPath = require("supertokens-node/lib/build/normalisedURLPath").default;
 let Session = require("supertokens-node/recipe/session");
 let express = require("express");
 let cookieParser = require("cookie-parser");
 let bodyParser = require("body-parser");
 let http = require("http");
 let cors = require("cors");
-let { startST, stopST, killAllST, setupST, cleanST, setKeyValueInConfig, maxVersion } = require("./utils");
+let {
+    startST,
+    stopST,
+    killAllST,
+    setupST,
+    cleanST,
+    setKeyValueInConfig,
+    maxVersion,
+    isProtectedPropName
+} = require("./utils");
 let { package_version } = require("../../lib/build/version");
 let { middleware, errorHandler } = require("supertokens-node/framework/express");
 let { verifySession } = require("supertokens-node/recipe/session/framework/express");
@@ -42,6 +53,53 @@ let lastSetEnableAntiCSRF = false;
 let lastSetEnableJWT = false;
 
 function getConfig(enableAntiCsrf, enableJWT, jwtPropertyName) {
+    if (enableJWT && maxVersion(supertokens_node_version, "14.0") === supertokens_node_version) {
+        return {
+            appInfo: {
+                appName: "SuperTokens",
+                apiDomain: "0.0.0.0:" + (process.env.NODE_PORT === undefined ? 8080 : process.env.NODE_PORT),
+                websiteDomain: "http://localhost.org:8080"
+            },
+            supertokens: {
+                connectionURI: "http://localhost:9000"
+            },
+            recipeList: [
+                Session.init({
+                    getTokenTransferMethod: process.env.TRANSFER_METHOD ? () => process.env.TRANSFER_METHOD : undefined,
+                    exposeAccessTokenToFrontendInCookieBasedAuth: true,
+                    errorHandlers: {
+                        onUnauthorised: (err, req, res) => {
+                            res.setStatusCode(401);
+                            res.sendJSONResponse({});
+                        }
+                    },
+                    antiCsrf: enableAntiCsrf ? "VIA_TOKEN" : "NONE",
+                    override: {
+                        apis: oI => {
+                            return {
+                                ...oI,
+                                refreshPOST: undefined
+                            };
+                        },
+                        functions: function (oI) {
+                            return {
+                                ...oI,
+                                createNewSession: async function (input) {
+                                    input.accessTokenPayload = {
+                                        ...input.accessTokenPayload,
+                                        customClaim: "customValue"
+                                    };
+
+                                    return await oI.createNewSession(input);
+                                }
+                            };
+                        }
+                    }
+                })
+            ]
+        };
+    }
+
     if (maxVersion(supertokens_node_version, "8.3") === supertokens_node_version && enableJWT) {
         return {
             appInfo: {
@@ -76,25 +134,13 @@ function getConfig(enableAntiCsrf, enableJWT, jwtPropertyName) {
                         functions: function (oI) {
                             return {
                                 ...oI,
-                                createNewSession: async function ({
-                                    req,
-                                    res,
-                                    userId,
-                                    accessTokenPayload,
-                                    sessionData
-                                }) {
-                                    accessTokenPayload = {
-                                        ...accessTokenPayload,
+                                createNewSession: async function (input) {
+                                    input.accessTokenPayload = {
+                                        ...input.accessTokenPayload,
                                         customClaim: "customValue"
                                     };
 
-                                    return await oI.createNewSession({
-                                        req,
-                                        res,
-                                        userId,
-                                        accessTokenPayload,
-                                        sessionData
-                                    });
+                                    return await oI.createNewSession(input);
                                 }
                             };
                         }
@@ -155,6 +201,41 @@ app.post("/login", async (req, res) => {
     res.send(session.getUserId());
 });
 
+app.post("/login-2.18", async (req, res) => {
+    // This CDI version is no longer supported by this SDK, but we want to ensure that sessions keep working after the upgrade
+    // We can hard-code the structure of the request&response, since this is a fixed CDI version and it's not going to change
+    Querier.apiVersion = "2.18";
+    const payload = req.body.payload || {};
+    const userId = req.body.userId;
+    const legacySessionResp = await Querier.getNewInstanceOrThrowError().sendPostRequest(
+        new NormalisedURLPath("/recipe/session"),
+        {
+            userId,
+            enableAntiCsrf: false,
+            userDataInJWT: payload,
+            userDataInDatabase: {}
+        }
+    );
+    Querier.apiVersion = undefined;
+
+    const legacyAccessToken = legacySessionResp.accessToken.token;
+    const legacyRefreshToken = legacySessionResp.refreshToken.token;
+
+    res.set("st-access-token", legacyAccessToken)
+        .set("st-refresh-token", legacyRefreshToken)
+        .set(
+            "front-token",
+            Buffer.from(
+                JSON.stringify({
+                    uid: userId,
+                    ate: Date.now() + 3600000,
+                    up: payload
+                })
+            ).toString("base64")
+        )
+        .send();
+});
+
 app.post("/startST", async (req, res) => {
     let accessTokenValidity = req.body.accessTokenValidity === undefined ? 1 : req.body.accessTokenValidity;
     let enableAntiCsrf = req.body.enableAntiCsrf === undefined ? true : req.body.enableAntiCsrf;
@@ -185,7 +266,8 @@ app.get("/featureFlags", async (req, res) => {
     res.status(200).json({
         sessionJwt:
             maxVersion(supertokens_node_version, "8.3") === supertokens_node_version && currentEnableJWT === true,
-        sessionClaims: maxVersion(supertokens_node_version, "12.0") === supertokens_node_version
+        sessionClaims: maxVersion(supertokens_node_version, "12.0") === supertokens_node_version,
+        v3AccessToken: maxVersion(supertokens_node_version, "14.0") === supertokens_node_version
     });
 });
 
@@ -269,8 +351,18 @@ app.post(
         if (req.session.getJWTPayload !== undefined) {
             await req.session.updateJWTPayload(req.body);
             res.json(req.session.getJWTPayload());
-        } else {
+        } else if (req.session.updateAccessTokenPayload) {
             await req.session.updateAccessTokenPayload(req.body);
+            res.json(req.session.getAccessTokenPayload());
+        } else {
+            let clearing = {};
+
+            for (const key of Object.keys(req.session.getAccessTokenPayload())) {
+                if (!isProtectedPropName(key)) {
+                    clearing[key] = null;
+                }
+            }
+            await req.session.mergeIntoAccessTokenPayload({ ...clearing, ...req.body });
             res.json(req.session.getAccessTokenPayload());
         }
     }
@@ -283,8 +375,18 @@ app.post(
         if (Session.getJWTPayload !== undefined) {
             await Session.updateJWTPayload(req.session.getHandle(), req.body);
             res.json(req.session.getJWTPayload());
-        } else {
+        } else if (Session.updateAccessTokenPayload) {
             await Session.updateAccessTokenPayload(req.session.getHandle(), req.body);
+            res.json(req.session.getAccessTokenPayload());
+        } else {
+            const info = await Session.getSessionInformation(req.session.getHandle());
+            let clearing = {};
+
+            for (const key of Object.keys(info.customClaimsInAccessTokenPayload)) {
+                clearing[key] = null;
+            }
+
+            await Session.mergeIntoAccessTokenPayload(req.session.getHandle(), { ...clearing, ...req.body });
             res.json(req.session.getAccessTokenPayload());
         }
     }