+++ test/utils.js
@@ -57,7 +57,7 @@ module.exports.getNumberOfTimesRefreshAttempted = async function (BASE = module.
 };
 
 module.exports.startST = async function (
-    accessTokenValidity = 1,
+    accessTokenValidity = 3,
     enableAntiCsrf = true,
     accessTokenSigningKeyUpdateInterval = undefined,
     enableJWT = undefined
@@ -131,6 +131,12 @@ module.exports.checkIfJWTIsEnabled = async function () {
     let featureFlags = await module.exports.getFeatureFlags();
 
     return featureFlags !== undefined && featureFlags !== null && featureFlags.sessionJwt === true;
+};
+
+module.exports.checkIfV3AccessTokenIsSupported = async function () {
+    let featureFlags = await module.exports.getFeatureFlags();
+
+    return featureFlags !== undefined && featureFlags !== null && featureFlags.v3AccessToken === true;
 };
 
 module.exports.checkSessionClaimsSupport = async function () {