+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [16.0.9] - 2023-05-05
 
+### Changes
+
+- Updated tests/test server to match new backend version supporting v3 access tokens
+
 ### Fixes
 
 - Fixes fetch interceptor calling clone on URL instances